LayoutTests/html5lib/runner-expected-html5.txt

CONSOLE MESSAGE: line 36: TypeError: Result of expression 'iframe.contentWindow.document.lastChild' [null] is not an object.
Content-Type: text/plain

#EOF
